Rearchitected changelog and defect resolution to be stored in json for more efficient dev
