+++ app.py
@@ -13,6 +13,7 @@ import datetime
 from flask import Flask, render_template, jsonify, request
 import firebase_admin
 from firebase_admin import credentials, firestore
+from version import APP_VERSION
 
 # --- Firebase Initialization ---
 # Check if the service account key file exists
@@ -35,7 +36,7 @@ print("✅ Firebase Initialized Successfully.")
 # This function fetches the version from Firestore, with a fallback.
 def get_app_version():
     # This version number will be incremented with each new set of changes.
-    default_version = '1.8.4'
+    default_version = APP_VERSION
     try:
         config_ref = db.collection('config').document('app_info')
         config_doc = config_ref.get()
@@ -737,6 +738,7 @@ def add_event(profile_id):
             'start_time': data.get('startTime'),
             'end_time': data.get('endTime'),
             'vehicles': data.get('vehicles', []),
+            'checklists': data.get('checklists', []),
             'is_raceday': data.get('isRaceday', False),
             'created_at': datetime.datetime.now(datetime.timezone.utc)
         }
@@ -788,6 +790,7 @@ def update_event(profile_id, event_id):
             'start_time': data.get('startTime'),
             'end_time': data.get('endTime'),
             'vehicles': data.get('vehicles', []),
+            'checklists': data.get('checklists', []),
             'is_raceday': data.get('isRaceday', False)
         }
         if not updates['name'] or not updates['start_time']:
@@ -812,12 +815,11 @@ def delete_event(profile_id, event_id):
         return jsonify({'success': False, 'message': f'An error occurred: {e}'}), 500
 
 
-@app.route('/get-next-raceday/<profile_id>', methods=['GET'])
-def get_next_raceday(profile_id):
+@app.route('/get-next-event/<profile_id>', methods=['GET'])
+def get_next_event(profile_id):
     try:
         now = datetime.datetime.now(datetime.timezone.utc)
         events_ref = db.collection('driver_profiles').document(profile_id).collection('events') \
-            .where('is_raceday', '==', True) \
             .where('start_time', '>=', now.isoformat()) \
             .order_by('start_time') \
             .limit(1) \
@@ -830,11 +832,11 @@ def get_next_raceday(profile_id):
         else:
             return jsonify({'success': True, 'event': None}), 200
     except Exception as e:
-        print(f"❌ Error getting next raceday: {e}")
+        print(f"❌ Error getting next event: {e}")
         return jsonify({'success': False, 'error': str(e)}), 500
 
 
-# NEW: Checklist Routes
+# --- Checklist Routes ---
 @app.route('/add-checklist/<profile_id>', methods=['POST'])
 def add_checklist(profile_id):
     try:
