+++ changelog.json
@@ -1,0 +1,39 @@
+[
+    {
+        "version": "1.8.7",
+        "date": "August 9, 2025",
+        "notes": [
+            "Architectural Improvement: Separated changelog and defect resolution data into their own JSON files for easier maintenance."
+        ]
+    },
+    {
+        "version": "1.8.6",
+        "date": "August 9, 2025",
+        "notes": [
+            "Architectural Improvement: Separated the app version into its own `version.py` file to simplify updates."
+        ]
+    },
+    {
+        "version": "1.8.5",
+        "date": "August 9, 2025",
+        "notes": [
+            "Fixed a UI defect where the user's helmet icon was not centered above the username in the profile selection modal."
+        ]
+    },
+    {
+        "version": "1.8.4",
+        "date": "August 9, 2025",
+        "notes": [
+            "Implemented the Advanced Race Day Checklists feature.",
+            "Added the ability to associate checklists with race events."
+        ]
+    },
+    {
+        "version": "1.8.3",
+        "date": "August 9, 2025",
+        "notes": [
+            "Added a prominent, global countdown timer for the next event.",
+            "Made the countdown timer a clickable element that navigates to the Race Schedule."
+        ]
+    }
+]
