+++ static/js/main.js
@@ -7,7 +7,7 @@ import { initAdmin } from './admin.js';
 import { initGarage } from './garage.js';
 import { initVehicle } from './vehicle.js';
 import { initSchedule, updateRacedayCountdown } from './schedule.js';
-import { initChecklists } from './checklist.js'; // NEW
+import { initChecklists } from './checklist.js';
 
 // --- Global App Object ---
 export const App = {
@@ -19,7 +19,7 @@ export const App = {
     loadGarages: null,
     loadVehicles: null,
     loadEvents: null,
-    loadChecklists: null, // NEW
+    loadChecklists: null,
 };
 
 // --- View Toggling Logic ---
@@ -33,7 +33,7 @@ const setView = (viewName) => {
     elements.garageManagementView.classList.add('hidden');
     elements.vehicleManagementView.classList.add('hidden');
     elements.raceScheduleView.classList.add('hidden');
-    elements.checklistManagementView.classList.add('hidden'); // NEW
+    elements.checklistManagementView.classList.add('hidden');
     elements.profileHeaderBtn.classList.add('hidden'); // Hide by default
 
     const isDevMode = viewName === 'developer';
@@ -68,7 +68,7 @@ const setView = (viewName) => {
     } else if (viewName === 'raceSchedule') {
         elements.raceScheduleView.classList.remove('hidden');
         if (App.loadEvents) App.loadEvents();
-    } else if (viewName === 'checklistManagement') { // NEW
+    } else if (viewName === 'checklistManagement') {
         elements.checklistManagementView.classList.remove('hidden');
         if (App.loadChecklists) App.loadChecklists();
     } else {
@@ -165,7 +165,7 @@ const initEventListeners = () => {
         setView('raceSchedule');
     });
 
-    elements.checklistTemplatesCard.addEventListener('click', () => { // NEW
+    elements.checklistTemplatesCard.addEventListener('click', () => {
         console.log("Click Event: 'Checklist Templates' card clicked.");
         setView('checklistManagement');
     });
@@ -215,5 +215,5 @@ document.addEventListener('DOMContentLoaded', () => {
     initGarage();
     initVehicle();
     initSchedule();
-    initChecklists(); // NEW
+    initChecklists();
 });
