+++ version.py
@@ -1,0 +1,2 @@
+# version.py
+APP_VERSION = '1.8.8'
